chore: organize global-styles in styles
